fix converstion

+++ src/value.c
@@ -29,7 +29,7 @@ static inline uint16_t binaryToBuffer (double binary, int base, char *buffer, ui
 			return snprintf(buffer, length, "%g", binary);
 		else
 		{
-			long dblDig10 = (long)pow(10, DBL_DIG);
+			double dblDig10 = pow(10, DBL_DIG);
 			int precision = binary >= -dblDig10 && binary <= dblDig10? DBL_DIG: 21;
 			
 			return snprintf(buffer, length, "%.*g", precision, binary);
